Add support for IBM i
There are systems where mount points do not have size definitions for
themselves. Therefor use the mapping between filesystem and storage
(hfFSStorageIndex) to get the storage device. Make sure, the device is
only added once. This makes it possible to use the plugin for IBM i
(formerly known as AS/400)