Plugin smart_: remove hard disk list weirdness
In 23a2e6645faa4493735836317119299c88e394dc (a broadly mixed merge
commit) the "hard_drive" variable was turned into a list (maybe related
to the "3ware RAID controller", but only its first component was used.
This change does not seem to have fulfilled any purpose up to now.

Maybe it was intended to change a name like "3ware-1" into "3ware/1",
but this is just wild guessing.

This change may result in one-time counter reset (due to the change of
the state file name) for "3ware RAID controller" disks. Besides this
there is no functional change.
Code readability is improved.