Nagios command format: ensure non-empty plugin output
Since 3f94d51a06 the fourth field of the lines written to "send_nsca"
could happen to be empty.  See the trivial example: "munin-limits --force".
This empty field prevented the receiver (nagios/icinga) from associating
this status report with a specific test (which is supposed to be encoded
in the fourth field of the output).

Thus we ensure, that "fofields" is always defined by falling back to
"ofields" (all fields that are currently in "OK" state).

Closes: #743, #1052