Enable evaluation of test results for test-common and test-plugins
Currently there are four sets of tests:
* test-common
* test-master
* test-node
* test-plugins

Out of these "test-master" and "test-node" currently fail for
non-obvious reasons.

The evaluation of the result (exitcode) of all tests was disabled in
2f57704157222a1ee9de35dd612a0d2e376a3174.

Since "test-common" and "test-plugin" work, we evaluate their result
and for now just ignore the results of the remaining known-to-fail
tests.