travis: ignore result of "test-master"
It seems to be impossible to use system-wide installed perl modules
(installed via apt) within the travis environment. Thus it would be
necessary to install RRD via cpan. Due to the lack of a suitable
package, it is not possible to run "make test-master" in the travis
environment. Thus we ignore its result for now.