Plugin mysql_: remove non-working (AREA|LINE)STACK backwards compatibility
Previously (before munin v1.3.3) stacked graphs required the first field
to use "AREA" or "LINE" and all following fields to use "STACK" for
"draw".
All following versions allow AREASTACK or LINESTACK for all attributes.
The plugin tried to work around the limitation of versions before
v1.3.4, but instead of using the field count for distinguishing the
first field from other fields, it used the "attribute in field" count.
Thus the plugins randomly emitted the "draw" attributes as "STACK",
"LINE" or "AREA" (due to the unordered nature of perl hashs).